nt_clockstuff.c: [[Fix NTP clock interpolation code]]
From: "Colin Dancer" <colin.dancer@pyrochrome.net>
To: <bugs@ntp.org>
Sent: 10 November 2000 12:59
Subject: NT bug in NTP 4.0.99j

I've found a bug in (and produced a fix for) the NT clock
interpolation code in NTP 4.0.99j.  The symptoms of the problem are
that gettimeofday() function on NT can be wrong by hundreds of seconds
if, while a gettimeofday() call is being processed, an APC completes
after the query of the performance counter but before the lock is
grabbed.  The most obvious fix is to move the lock to include the
querying of the performance counter, but this could affect the
predictability of the timestamp so I have instead tweaked the code to
detect and sidestep the duff calculation.  I've also found that on a
loaded system the execution of the APC can be delayed, leading to
errors of upto 10ms.  There is no easy fix to this, but I do have code
for an alternative interpolation scheme which avoids the problem on
single processor systems. I'm currently integrating this along with
code for deciding which algorithm to use based on whether the system
is SP or MP.