machines.c:  From: Ron Thornton <rthornto@pictel.com>
In libntp/machines.c, the getpass() function is implemented for
SYS_WINNT using the _getch() call.  On my NT 4.0 systems, the enter
key does not return '\n' so the password read does not end.  To work
around the problem, I must enter Ctrl-J (linefeed) to terminate the
password entry.  To fix the problem, I had to change
"((c=_getch())!='\n')" to "((c=_getch())!='\n' && c!='\r')" in
machines.c I left the test for \n (linefeed) in the event microsoft
alters _getch() in the future and added the test for \r (enter/carrage
return) to cope with the current implementation of _getch()