[[Fix bug when signaled I/O is used]]
This patch fixes a bug in NTP which appears if NTP is built on a
system where signaled I/O is used, i.e. HAVE_SIGNALED_IO is defined in
config.h. The bug has been found in ntp-4.1.71 but seems to last in
the NTP code since the xntp3 days.

The symptom: If a radio clock is connected via the serial port using
the parse driver (driver 8) then there are lots of messages in the
syslog saying:

> 31 Oct 09:13:42 ntpd[6708]: input_handler: handler_count is 2!
> 31 Oct 09:13:42 ntpd[6708]: input_handler: handler_count is 2!

This has been observed under AIX 4.3.2 and Solaris 8, and the patch
works in both cases. This effect is very obvious if a radio clock is
connected via a serial interface, and a SIGIO signal is raised
whenever a character has been received. It's a common problem,
however, but the probability that a signal is raised again while the
handler for the previous signal is still executing is much less if
there are only signals if network packets are received.

The fix: The patch below is used to take care that the SIGIO signal is
not unintentionally unblocked inside the sigio_handler() if the
handler executes a piece of code that is bracketed by
BLOCKIO()/UNBLOCKIO() calls.