Second try at getting the refid stuff right for ipv6.
Add an addr_refid fieldto the interface structure and hash each
interface address on startup. This isthen used to compare the incoming
refid's against. The hash function,addr2refid() use md5 on the ipv6
addresses and use the ipv4 addresses just asit is. Add sys_peer_refid
which is the hash value of the ip address of sys_peerand gets
recalculated in clock_select() when sys_peer change. This way
thebehaviour for the ipv4 case stay the same as previous
implementations.