For BSDs, use setrlimit() to limit the stacksize...
...to something considerablysmaller than the default, similarly to
what's done on AIX. This was culledfrom the NetBSD sources. Christos
Zoulas gave credit for the patch, in thecommit log, to Witold J. Wnuk
(against ntp-4.0.99i, 2002-04-18). I've addeda configure-time check
for setrlimit(), and I had to up the pages from "8"to "20" to keep
from dumping core in the resolver. (Apparently, the resolvernow uses
"a lot" of stack space to initialize, relatively speaking: thedefault
is something like "512", but p_vm_ssize is only about "5" with
therunning program at any moment I look at it on both of the hosts I
tested.)The bottom line is that it reduces the resident set size by
about 2 Mb, from4.5 Mb, or by about 44%.