Changes from Dave Mills:
1. Servers presented with key IDs they don't have (a common
mistake) return a crypt-NAK. This also happens when the server rolls
new private values daily without telling the clients. To protect
against bad guys tossing an old duplicate over the wall, the client
doesn't believe this unless the timestamps pass the loopback check. If
the server was previously reachable, the client simply restarts the
association without missing a beat. If not, and to protect against a
restart loop or later server restart, the client disregards the NAK and
continues (while disabling burst mode). If really invalid, the
association timeout will eventually expire following which a persistent
association is marked permanently disabled or an ephemeral association
is demobilized.

2. One of the most common mistakes is to goof up the
identity credentials. If the client received these credentials via
mail, the script in the message installs the file and link
automatically. If done by hand and done wrong, either the server or
client will find the error and disable the association. As in all such
cases, a server or client ephemeral association is demobilized and the
resources recovered.

3. Symmetric modes have always been tricky, much more so with
Autokey,and even more so with misconfigured identity keys. It gets
even wilderwhen either peer is restarted in the middle of
negotiations. All the evil scenarios I can think of seem to come out
on the other side,although there are a few that have to wait the
association timeout toproperly close down and recover resources.

4. The association is permanently disabled upon kiss-o'-death or
unrecoverable crypto error. So, I re-enabled log reporting for
these smooches. At least the offender is notified.

5. When a really tenacious clogger shows up and lights the call-gap,
the server puts it in the restrict list. He stays there until the
server is restarted. Not only does this discourage the terriot, but
evidence is collected for possible further action.

What more do we have to do to turn this stuff into a real Secure
Timestamping Service? I think the security model, protocols, abuse
hardening and time performance are consistent with a bulletproof
service assuming, that is, PGP mail is used to get the keys and the
servercertificates are properly signed.