[[Correct 12-channel OnCore behavior]]
ntpd/refclock_oncore.c: The oncore hardware provides a 'negative
saw-tooth' the offset of the PPS signal which is on the zero crossing
of its internal clock, from its true value.  the refclock_oncore.c
driver applies this difference to correct the time of the provided
timestamp.  This value was being applied correctly for the 6 and 8
channel receivers, but not for the 12 channel.  This has been
corrected.