This is a group of 4 patches that
(1) replace blanks with tabs in the source.

(2) provide a cleaner solution to turning FLAG_PPS on/off in
oncore_get_timestamp.

(3) change LOG_INFO -> LOG_DEFAULT for one message, reducing the
extraneous message in the messages log.

(4) an implementation of [Bug1390] initially provided by Russell
Yount, to turn the PPS signal on/off.  Note that this does NOT affect
timekeeping.