Fix misleading indent - Coverity CID 117238.
This patch might be wrong - original coder's intent might have been
to send BJ, *then* Bl in June and December, in which case the right
thing would be to wrap both statements in braces.  I'm not too
worried, as this is a corner case in a driver for a long discontinued
device, to cover for buggy firmware that was obsolete when the driver
was written.