Change name and ID of Spectracom driver.  No logic changes.
Type 4 is no longer for a WWVB radio and has not been since before the
WWVB modulation change in 2013. Now it handles Spectracom 9xxx and
SecureSync GPS receivers.

Accordingly, the refid needed to be changed from WWVB to GPS and many references
to "Spectracom WWVB" in documentation and comments fixed.