Revert overzealous deletion of some time-library code.
This could have caused build failure on a 32-bit Unix without timegm(),
e.g. not using libc6.

There's a check that will throw an error if timegm() is absent on a
machine with 64-bit time.  I dom't know why it's there, as the code
looks 64-bit clean - overflow issues on extreme values, perhaps?  This
will need to be troubleshot someday.