Refactor so ntp_config no longer needs to know about option processing.
A first step in disentangling autogen from the rest of the code.

Note: this only fixes a symptom, by turning the option-parse calls into
evaluations of booleans set by those same option-parse calls just before
config_ntpd() is called.  The real problem is poor separation of concerns
- the actions these options are controlling shouldn't be done inside
configuration parsing at all.  Someday that will have to be fixed.

This change depends on the assumption that a config via ntpq never
sets interface or saveconfigquit options.  This *should* certainly
be the case, otherwise ntpq could be used for remote DoS.