Pull the second stage of option processing out of getconfig().
It previously lived inside getconfig() because it needed to be after
the computation of the default config file location. I fixed that by
adding an ugly global, alas.  This whole sequence will vet a serious
refactor someday but right now nminimal change is a goal.

This was a step towards consolidating all option processing in one place.

Tested by successfully invoking saveconfigquit.