Remove an autoconf dependency, on SIZEOF_SHORT.
Because it's really annoying to have compilation of a file that
includes ntp_types crap out because config.h wasn't also included.

The #define itself should no longer be needed, as we're assuming we
have ANSI fixed-length types when we need them.  However, removing
this check creates the possibility of a subtle problem near
WORDS_BIGENDIAN.

That is, if the compiler doesn't have int64_t (and thus INT64_MAX)
accesses to vints on big-endian machines may use the little-endian
structure layout.  This should not be a problem as long as vint
storage is never cast to an int64, which currently is the case.