Assume we have 64-bit scalar types int64_t and uint64_t.
In ancient times the core code didn't assume this, but the dependency snuck
back in with the ISC library and libevent.  So, we shrug and drop a whole
bunch of fragile code that probably hasn't even been compiled in a decade.

Note, this doesn't git rid of vint64.  That will be much more difficult
because of the arcane way the calendar code is written. It's a step
in that direction, though.