Remove driver 12, the KSI/Odetics TPRO/S IRIG Interface (IRIG_TPRO)
This was a clock source that connected directly to the SBus of a Sun
workstation. SBus was replaced rapidly by PCI after 1997; no hardware
based on SBus has been shipped for at least a decade and the last
of these were probably shipped before the year 2000.

It won't build without a Sun-specific header tied to an obsolete SunOS
version.  It is quite likely nobody has actually compiled this code
in more than a decade.