Fix a compiler warning about strptime(3).
The way to fix this, making the prototype visible, turns out to disable
some typedefs that are normally declared in the GCC version of sys/types.h.
The workaround is to declarte them explicitly in ntp.h.

This fix also made visible some uses of the obsolete BSDism caddr_t,
which have been removed.