Fix internal header dependencies.
I havn't hit this before in previous projects.  I was appending to INCPATHS when
I should have been appending to includes for a target.  This ensures the
dependencies are picked up properly.

For some reason, this breaks the build I don't know why I will be following up
with another commit to fix breakage.