Removed driver 43: RIPE NCC interface for Trimble Palisade
The code for this driver cannot be salvaged in anything like its
present form without using unsafe string functions, so it gets
removed. This shouldn't be a problem as it was written for a specially
customized device in an experimental program that was shut down in
2014.

If anyone wants it back, we'll toss them the driver source from the
repository history and say "fix it". It seems highly unlikely this
will occur.