Conversion of the Mills HTML docs to asciidoc lands with this commit.
The intent is (a) to make these easier to maintain, and (b) to have just one
master format for all in-tree documentation.  The latter is not yet completely
accomplished, as various manual pages in the source directories are not
yet converted.

The conversion resembles the ancestral HTML pretty closely, using an
adaptation of Dave Mills's stylesheets to retain his look.  The
content has been verified to match by two reviewers.  The sppearance
is very similar, though not identical; asciidoc doesn't allow quite as
much freedom in image placement as raw HTML, and table boxes don't
look quite the same. The look of the command.txt and footer.txt
includes isn't completely right either.