Drop the PSTI/Traconex WWV/WWVH Receiver (type 3)
This driver cannot have worked since the WWVB modulation change in 2013.

This device is described here: http://prc68.com/I/PST1020.shtml with a
2005 note that it is "not very reliable" having drifted by a full minute.

Investigating the vendor site at U.S. Traffic Corp (now Peek Traffic)
reveals that this device was discontinued, probably before 2000.

According to this August 2001 post

https://groups.google.com/forum/#!topic/comp.protocols.time.ntp/0jOItglBTLE

the device required manual intervention every year and was not
designed for use after 2001.  Dave Mills replied to that one in a way
that implies he still has two of these and there is a workaround to
the problem described, but also says they have a service life of only
a few years due to a failure-prone component.