Document-namespace maneuvers.
In the new scheme, docs/ and external manpages share nost of their contents
by including a body file. This way there is just one point of truth to keep
updated about any given feature.

A tricky point is that while asciidoc includes work in the docs/ versions,
they clash with the way asciidoc processes documents of doctype manpage.
So we need to template in the inclusions before asciidoc sees them.

This is why the manual-page wrappers have *.txt.in extensions, to mark
that they need to be preprocessed into *.txt files by expanding file
inclusions wrapped in @ @.