Get rid of ISC_MUTEX_PROFILE.
This is a long step towards replacing the ISC thread/mutex stuff with
bare pthreads.  If we're assuming ANSI pthreads the ISC layer is just
unnecessary bulk, and the right way to do profiling is to link a
profiling version of the pthreads library.

Also clean up some guard symbols and a few unnecessary macros.