Rewrite the signaled-IO code to be POSIX-conformant.
Contrary to what some comments in the configure.ac for NTP classic implied,
this would probably work under Linux and *BSD.

I have not enabled this code yet because asynchronous I/O is
notoriously difficult to debug and we don't need that complication
working towards a first release.  The main purpose of this rewrite
was to throw out a lot of dubious backwards-compatibility cruft.

Later, if we find we need the performance boost, we can enable and test this.
If we don't need the performance the path of wisdom would be to throw it out.