Remove support for the Symmetricom bc600-series GPS/IRIG Receiver.
The crash landing here is that it requires a closed-source driver
with kernel privileges.  Mark Atwood and I have judged this to be
an unacceptable security risk.

This commit also adds a note to the docs/refclocks.txt explaining
that as a policy matter refclocks requiring privileged closed-source
blobs will not be supported.