A major step towards getting the configuration documentation refactored.
What this commit does is move most of the material on authentication and
crypto from ntp.conf(5) to the Authentication page in the web documentation.

This material didn't belong on the man page, which should mainly be an
option reference, and it was actually better exposition than what was
already there.  What used to be there has been slimmed down and is
now the sections describing Algorithms and Data Formats.

Some duplication between new and old content remains to be polished out
by future edits.

Also, describe update-leap in scripts/README.