Documentation header, footer, and include visual cleanup.
Fix up the title header on the index page.

Get rid of command.txt include, it duplicated links that were always
present in the footer.

Tweak the footer to look a bit more like the original Mills HTML.
Not a perfect match but as good as we're likely to get.

Fix a stale ntptickadj reference to point to ntpfrob.

Repair some broken inclusions.