Remove detection and inclusion of netinet/in_systm.h and netinet/ip.h.
These are not in the Open Group standard, and not required under Linux
or FreeBSD.  No code was conditioned on HAVE_NETINET_IP_H or
HAVE_NETINET_IN_SYSTM_H, just the inclusions.  So if this breaks
anyting on any target platform it should cause simple compile-time
errors due to missing definitions.

Google searches hint that these may once have been required on older
Unixes, but the definitions we require have been absorbed into the
standardized netinet/in.h header.  Note however that this change has
*not* been tested under Mac OS X.