License exception cleanup in libntp/ntp_rfc2553.c
Used to contain two copies of the BSD license - one 3-clause variant
attached to a copyright by the WIDE project and one 4-clause variant
with a Regents of the University of California copyright.

The 2003 annulment of the 4th clause makes the second license the same
as the first. Tag appropriately.

Approved-by: Mark Atwood <me@mark.atwood.name>