Attempt to fix signed-vs.-unsigned issues revealed by -Wextra.
The theme in this patch is changing variables that are semantically data
lengths from int to size_t.  This cleanly resolves many sign-compare bugs.
It requires some printfs to change from %d to %zd formats.

No logic changes. Associated code has been eyeball-checked for subtractions that
might induce modular wraparound, but if a future core dump produces a bisection
that lands here that will be the thing to suspect.