Add support for a 'Release mode'
Steps to use:
  1. Build as normal with manpages enabled.
  2. Edit wscript and change NTPS_RELEASE to True
  3. run 'make dist'

This will do a two things:

  1. Remove dependency on Bison and include generated files
  2. Include prebuilt manpages and remove dependency on asciidoc

This is most important for #1 as we can ship our pretested generated files
removing users using the tarball from having a buggy version of Bison.

Having prebuilt manpages is good for those who don't want to deal with asciidoc
on smaller systems.  It also makes things easier for packaging on some systems.