Add a note about libevent on some systems.
On NetBSD there is no event_core.  event is pulled in by event_pthread so
event_core is not detected however it builds anyway since the link test runs.

This is by design as a 'fall through' to detect strange setups.  Still not sure
how to handle this case yet for now users will get a harmless warning.