Drop support in ntpfrob for ancient RS6000 bug
According to https://www.eecis.udel.edu/~mills/ntp/html/hints/rs6000 ,
the a hack was added to xntp in v2 for AIX 3.1, and that was before
"93/12/3" , but this version would not work on older AIX.

AIX 3.1 was released in Feb 1990, so the bug was in something older than that.
AIX is still supported, and is 7.2 currently.