Perform all bogon checks before updating peer variables (CVE-2016-4954)
NTP Classic bug #3044. Although NTP Classic assigned this a CVE, I
don't regard it a security issue because the tardy checks aren't
relevant to authentication. Failing to sanity-check the leap, stratum,
and root dispersion fields doesn't make it any easier for an adversary
to spoof responses, and it doesn't enable an authentic-but-malicious
server to put the client into any state worse than one that it could
put it into by supplying values that pass these checks.