Remove dead code.
The third loop on i that was here before this commit - this:

	for (; i < ipbytes; i++) {
		if (p[i] != 0)
			return (ISC_R_MASKNONCONTIG);
		i++;
	}

could never be reached, because at the end of the second loop it is
necessarily true that either an error return has occurred or i >= ipbytes.

Harmless bug, but the dead code caused a compiler warning under clang.
Mentioned in GitLab issue #71.