Greatly reduce the amount of code dependent on knowing magic address bits.
Instead of dispatching on the type fields of magic refclock addresses
to get the type name of a driver by indexing into a separate table, go
from the peer to the description in the driver method table.

With this change, very little code in ntpd itself still knows about
the magic 127.127.t.u addresses used to distinguish refclocks in
ntp.conf, and the few remaining dependencies (find them by grepping
for ISREFCLOCKADR) will be easily removable later. The uses in ntpq
will be more difficult to drop.