Add a basename element to the driver transfer vector.
Needed for the new  refclock syntax. Yes, it's a little strange that the
per clockname member is initialized at driver sartup rather than from this,
but the distinction is required for the parse driver. At config time it
will be known as "generic", but wants to log by its subtype.