Documentation mega-patch to go with the new refclock syntax.
Replace almost all clock configuration examples in documentation and comments
with equivalents in the new syntax. (The few exceptions are historically
marked.)

Change all uses of the term 'fudge' with respect to flag bits. These
are now described as "driver options". The term 'fudge' is reserved
for time fudges.

Drop the old configuration examples from the UDel machines.

Add a note to the documentation start page about the new syntax,
mentioning that the old one is still supported.