Do not attempt to lock ntpd into memory.
This had a bad interaction with the async-DNS code that leads to
sporadic crashes. There are some possible fixes that add code
complexity in Classic, but the right thing to do is (a) notice that
this call is pretty useless on modern systems, which swap so seldom
that many people don't bother with swap partitions, and (b) simplify
by dropping the code.

One good reason to retain this might have been crypto security.  But
our crypto geek says: "Some people like to use memory locking to make
sure that intermediate results of crypto computations don't get
written to swap and sit there long-term. But that's only useful if
you're not already storing keys on disk (rather than on, say, an HSM),
and we don't support any alternative to that."