Refactor receive() and process_packet()
This is the first and biggest step of a complete rewrite of NTP's
receive-side logic. receive() now uses a struct parsed_pkt instead of
struct pkt, and the function is much better organized. Authentication
is now performed as early as possible rather than as late as possible;
this isn't the 1980s any more and checking a MAC is no longer
expensive. Other security-relevant checks such as orgin timestamp
validation are also moved to the fore.

Notable bug fixes introduced by this commit:

* Server packets are now immediately rejected as bogus if there is no
  request outstanding (peer->outcount == 0).

* The old code was sticking the packet's transmit timestamp into the
  peer structure's receive timestamp, a bug that appears to have been
  introduced in 2008 when interleaved mode was implemented. This is
  likely to have been harming timekeeping precision all this years,
  and the fix may be an especially big win for clients which talk to
  a stratum 1 server via a LAN.

* KoDs are handled more sensibly and will never bump the polling
  interval up to anything ridiculous.

Known issues as of this commit:

* Only basic client and server modes are currently working. Support
  for the myriad other association modes will be reintroduced in
  later commits.

* MS-SNTP support is not yet re-implemented.

* The call to record_raw_stats() is missing.

* Some system counters such as ntp_oldversion and ntp_newversion
  are not incremented at appropriate times.

* There are no logging calls in the new code.

Current testing status: I've verified that the code basically works in
the common cases both as a client and as a server. I haven't tested
authentication yet, and haven't verified that every sort of bogon is
correctly rejected.