More magic-address banishment.
This commit is a bit of a grab-bag.  First, it adds a build option
--enable-classic-mode that can be used to prevent the new behaviors
the rest of the patch implements.

Then, it modifies some logging functions so that the
clock-identification fields in them become composed from the driver
shortname string and the unit number, rather than being a 127.127.t.u
address.  The --enable-classic-mode option *restores* the old behavior.

Finally, it modifies the documentation to reflect the new behavior.