Revert "Terminology change: mode -> subtype when it refers to a device type."
Adding "subtype" to the token table as an alias for "mode" apparently confuses
the crap out of the scanner generator.  The symptom of this is strangely
specific - the "peerstats" keyword is no longer recognized, breaking logging.

This reversion restores things to a good state.  Next, figure out out why
the scanner generator is losing its merbles.