First try at activating seccomp
seccomp restricts the syscalls a program can do.
There was code in ntp classic, but it didn't work.
It needs a library.  (I assume the code was all in
the kernel a long time ago.)

Added to waf configure
  --enable-early-droproot
  --enable-seccomp

Without either, there shouldn't be any changes.

Early droproot requires that user ntp can open all the files.
There are two known case (so far) that don't work.
One is SHM - the first 2 slots are restricted to root.
The other is opening sockets on NetBSD.  I assume it's
a check for system ports (< 1024 or such), but I haven't
checked.

Enable seccomp works for all my test cases, but there
could easily be obscure cases I haven't tried.  We will
need a run time flag to disable it before we make the
default to be on.  The system calls are an interesting
tangle.  The actual call that gets used by libc depends
on 32 vs 64 and the age of the program vs the age of the
kernel.

Seccomp is Linux only.  It doesn't work on ARM: Invalid argument