Allow minpoll to be set as low as 0.  Useful on fast networks.
chrony allows minpoll 0 and doesn't blow up the net, so this is
probably safe.  A higher lower limit made sense when WAN capacity was
much lower and we didn't want aggressive configurations grabbing a
lot of it. But today the worst NTP can do is spit in the ocean
comparesd to, say video streaming traffic.

Note that this does not change the default minpoll value, just the
lowest you can explicitly set.  The documentation lied about that,
claiming it was 4 when it had been 3 since an unexplained change in
2008.