snap: provide and consume config in user accessible place
To make the config more than just a default it has to be made available
in the $SNAP_DATA directory. This is a system config, so it didn't
qualify for $SNAP_USER_DATA.

Then on the other hand ntpd has to be made to consume that configuration
and at the same time avoid https://gitlab.com/NTPsec/ntpsec/issues/98.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>