snapcraft: update known TODOs
conffile access is avaiable now. Further improvements can be tackled
later on once:
- https://gitlab.com/NTPsec/ntpsec/issues/98
- https://bugs.launchpad.net/snapcraft/+bug/1611287
are solved. Both bugs are linked at the respective code, so we can drop
the generic "make config available" as we have a way to do it.

Also indent the snap isolation related tasks under the top "fix
isolation" task.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>