Remove unused support for using signaled I/O.
This was a fallback for operating systems that don't have
packet-receive timestamps as a built-in feature of the network layer.
But these systems have another fallback - using the system clock time
right after arrival in userland - and logic in the the NTP Classic
build recipe suggests signaled I/O didn't work on Linux or *BSD or a
large variety of other Unixes clear back to stock System V.

Accordingly, drop this odead code out to reduce complexity.