Remove the (broken) trap feature.
This feature was broken in NTP Classic before the NTPsec fork; traps
were configured before interfaces are configured causing any trap
commands to fail to find the appropriate interface.  Reported as
Classic bug #2923 and NTPsec issue #106.

If this feature still worked, even slight bugginess in it could lead
to DoS vulnerabilities.  There is, however, reason to suspect it has
been broken for a long time, with web searches finding no evidence of
even exploratory use since 2007.

Thus, the right thing to do is drop it. In the future something equivalent
might be implemented through something like an SNMP service, if there is
user demand.